move serializeValue() to serializer for table serializer
