+++ hugegraph-core/src/main/java/com/baidu/hugegraph/backend/serializer/TableSerializer.java
@@ -19,7 +19,9 @@
 
 package com.baidu.hugegraph.backend.serializer;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 import java.util.Set;
 
 import com.baidu.hugegraph.HugeGraph;
@@ -353,16 +355,27 @@ public abstract class TableSerializer extends AbstractSerializer {
 
     @Override
     protected Query writeQueryCondition(Query query) {
-        if (query.resultType().isGraph()) {
-            ConditionQuery result = (ConditionQuery) query;
-            // No user-prop when serialize
-            assert result.allSysprop();
-            for (Condition.Relation r : result.relations()) {
-                if (r.relation() == Condition.RelationType.CONTAINS) {
-                    r.serialValue(JsonUtil.toJson(r.value()));
+        ConditionQuery result = (ConditionQuery) query;
+        // No user-prop when serialize
+        assert result.allSysprop();
+        for (Condition.Relation r : result.relations()) {
+            if (r.relation() == Condition.RelationType.IN) {
+                List<?> values = (List<?>) r.value();
+                List<Object> serializedValues = new ArrayList<>(values.size());
+                for (Object v : values) {
+                    serializedValues.add(this.serializeValue(v));
                 }
+                r.serialValue(serializedValues);
+            } else {
+                r.serialValue(this.serializeValue(r.value()));
+            }
+
+            if (query.resultType().isGraph() &&
+                r.relation() == Condition.RelationType.CONTAINS) {
+                r.serialValue(JsonUtil.toJson(r.serialValue()));
             }
         }
+
         return query;
     }
 
@@ -576,6 +589,20 @@ public abstract class TableSerializer extends AbstractSerializer {
 
     protected abstract void parseProperties(HugeElement element,
                                             TableBackendEntry.Row row);
+
+    protected Object serializeValue(Object value) {
+        if (value instanceof Id) {
+            value = ((Id) value).asObject();
+        }
+        if (value instanceof String) {
+            value = escapeStrings((String) value);
+        }
+        return value;
+    }
+
+    protected String escapeStrings(String value) {
+        return value;
+    }
 
     protected void writeEnableLabelIndex(SchemaLabel schema,
                                          TableBackendEntry entry) {
