+++ hugegraph-mysql/src/main/java/com/baidu/hugegraph/backend/store/mysql/MysqlSerializer.java
@@ -162,4 +162,9 @@ public class MysqlSerializer extends TableSerializer {
             schema.userdata(e.getKey(), e.getValue());
         }
     }
+
+    @Override
+    protected String escapeStrings(String value) {
+        return MysqlUtil.escapeString(value);
+    }
 }
