+++ hugegraph-postgresql/src/main/java/com/baidu/hugegraph/backend/store/postgresql/PostgresqlSerializer.java
@@ -19,8 +19,12 @@
 
 package com.baidu.hugegraph.backend.store.postgresql;
 
-import org.apache.logging.log4j.util.Strings;
+import java.sql.SQLException;
 
+import org.apache.logging.log4j.util.Strings;
+import org.postgresql.core.Utils;
+
+import com.baidu.hugegraph.HugeException;
 import com.baidu.hugegraph.backend.id.IdUtil;
 import com.baidu.hugegraph.backend.serializer.TableBackendEntry;
 import com.baidu.hugegraph.backend.store.BackendEntry;
@@ -51,5 +55,21 @@ public class PostgresqlSerializer extends MysqlSerializer {
             entry.subId(index.elementId());
         }
         return entry;
+    }
+
+    @Override
+    protected String escapeStrings(String value) {
+        if (value.equals("\u0000")) {
+            return "\'\'";
+        }
+        StringBuilder builder = new StringBuilder(32);
+        builder.append('\'');
+        try {
+            Utils.escapeLiteral(builder, value, false);
+        } catch (SQLException e) {
+            throw new HugeException("Failed to escape '%s'", e, value);
+        }
+        builder.append('\'');
+        return builder.toString();
     }
 }
